chore: rename file to class name
