+++ curation-migration-datasync/dynamodb/curatedItemRecordModel.ts
